finished behodler Admin screen

+++ client/src/blockchain/contractInterfaces/ERC20.ts
@@ -5,7 +5,7 @@ export interface ERC20 extends BaseContract {
 	totalSupply: () => any
 	balanceOf: (address: address) => any
 	allowance: (owner: address, spender: address) => any
-	transfer: (to: address, value: Uint16Array) => any
+	transfer: (to: address, value: uint) => any
 	approve: (spender: address, value: uint) => any
 	transferFrom: (from: address, to: address, value: uint) => any
 	increaseAllowance: (spender: address, addedValue: uint) => any
